requirements.txt
# EduVidQA — Combined requirements
# Install with: pip install -r requirements.txt

# ── Backend (FastAPI) ────────────────────────────────────────────
fastapi==0.135.2
uvicorn[standard]==0.42.0
pydantic==2.12.5
python-multipart
python-dotenv
eval_type_backport  # Python 3.9 compat for X|None syntax in Pydantic

# ── Session A: Ingestion ────────────────────────────────────────
youtube-transcript-api
yt-dlp==2026.3.17
ffmpeg-python
Pillow==12.1.1

# ── Session B: RAG ──────────────────────────────────────────────
sentence-transformers==5.3.0
# chromadb removed in H2 — replaced by Supabase pgvector

# ── Session C: Inference ────────────────────────────────────────
transformers==5.4.0
bitsandbytes
accelerate
qwen-vl-utils
torch==2.11.0
huggingface_hub

# ── LLM API clients (answer.py) ─────────────────────────────────
groq==1.1.2
google-genai==1.69.0

# ── Session G: Supabase ─────────────────────────────────────────
supabase
psycopg2-binary

# ── Session I: Auth + Rate Limiting ─────────────────────────────
PyJWT
slowapi

# ── Session L: Observability ────────────────────────────────────
sentry-sdk[fastapi]

# ── Testing ─────────────────────────────────────────────────────
httpx
pytest
pytest-asyncio
